RefactorServer/requirements.txt
flask==3.0.3
flask-session==0.8.0
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-swagger-ui==4.11.1
supabase==2.4.1
nltk==3.8.1
pyvi==0.1.1
gensim==3.8.0
scikit-learn==1.4.0
scipy==1.12.0
stripe==9.6.0
pytest==8.2.1
python-dotenv==1.0.1
gunicorn==22.0.0
networkx==3.2.1